rest cleanup, generate view/rest for device using yeoman

+++ src/main/java/pl/lodz/p/repository/DeviceRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import pl.lodz.p.domain.entities.Device;
 
-public interface DeviceRepository extends JpaRepository<Device, Integer> {
+public interface DeviceRepository extends JpaRepository<Device, Long> {
 
 }
